+++ src/config/index.ts
@@ -1,0 +1,4 @@
+export * from './app.config';
+export * from './i18n.config';
+export * from './swagger.config';
+export * from './typeorm.config';
