+++ src/modules/movies/movies.service.ts
@@ -3,8 +3,7 @@ import { HttpStatus, Injectable, Logger } from '@nestjs/common';
 import { CreateMovieDto } from './dto/create-movie.dto';
 import { UpdateMovieDto } from './dto/update-movie.dto';
 import { FindAllMoviesDto } from './dto/findAll-movies.dto';
-import { calculatePagination } from '@/utils/getSkipPage';
-import { calculatePaginationResponse } from '@/utils/calculatePaginationResponse';
+import { calculatePagination, calculatePaginationResponse } from '@/utils';
 import { I18nException } from '@/common/exceptions/i18n.exception';
 import { FindOptionsWhere, ILike, Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
