+++ src/modules/scenes/scenes.service.ts
@@ -5,8 +5,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Scene } from './entities/scene.entity';
 import { FindAllSceneDto } from './dto/findAll-scene.dto';
-import { calculatePagination } from '@/utils/getSkipPage';
-import { calculatePaginationResponse } from '@/utils/calculatePaginationResponse';
+import { calculatePagination, calculatePaginationResponse } from '@/utils';
 import { I18nException } from '@/common/exceptions/i18n.exception';
 import { Movie } from '../movies/entities/movie.entity';
 import { I18nService } from 'nestjs-i18n';
