+++ src/modules/tracks/tracks.service.ts
@@ -5,8 +5,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { FindOptionsWhere, Repository } from 'typeorm';
 import { Track } from './entities/track.entity';
 import { FindAllByMovieIdTrackDto } from './dto/findAllByMovieID-track.dto';
-import { calculatePagination } from '@/utils/getSkipPage';
-import { calculatePaginationResponse } from '@/utils/calculatePaginationResponse';
+import { calculatePagination, calculatePaginationResponse } from '@/utils';
 import { FindAllBySceneIdTrackDto } from './dto/findAllBySceneID-track.dto';
 import { Song } from '../songs/entities/song.entity';
 import { Scene } from '../scenes/entities/scene.entity';
